# functions/swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Bet Chicago
host: localhost:5001
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
  - application/octet-stream
  - application/vnd.contentful.management.v1+json
  - application/x-www-form-urlencoded
  - multipart/form-data
  - text/plain; charset=utf-8
  - image/jpeg
produces:
  - application/json
paths:
...
  /nfl/updateschedule:
    x-swagger-router-controller: nfl
    post:
      description: update schedule
      operationId: updateSchedule
      tags:
        - NFL
      parameters:
        - name: year
          in: query
          description: year 2018, 2017, ... 2012 - maybe farther
          required: false
          type: string
        - name: tag
          in: query
          description: pre, regular, post
          required: true
          type: string
          enum:
            - REG
            - PRE
            - PST
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /nfl/updatestandings:
    x-swagger-router-controller: nfl
    post:
      description: update standings
      operationId: updateStandings
      tags:
        - NFL
      parameters:
        - name: year
          in: query
          description: year 2018, 2017, ... 2012 - maybe farther
          required: false
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /nfl/updateweekly:
    x-swagger-router-controller: nfl
    post:
      description: update weekly schedule
      operationId: updateWeekly
      tags:
        - NFL
      parameters:
        - name: year
          in: query
          description: year 2018, 2017, ... 2012 - maybe farther
          required: false
          type: string
        - name: tag
          in: query
          description: pre, regular, post
          required: true
          type: string
          enum:
            - REG
            - PRE
            - PST
        - name: week
          in: query
          description: 1-4, 1-17, 1-4 (pre/reg/pst)
          required: true
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
...
  /nfl/updateplayerpage:
    x-swagger-router-controller: nfl
    post:
      description: update player stats for a page of ids
      operationId: updatePlayerPage
      tags:
        - NFL
      parameters:
        - name: skip
          in: query
          description: start pos in list
          required: false
          type: string
        - name: pagesize
          in: query
          description: player count to update (20 default)
          required: false
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /nfl/updateplayer:
    x-swagger-router-controller: nfl
    post:
      description: update player profile/stats
      operationId: updatePlayerProfile
      tags:
        - NFL
      parameters:
        - name: playerid
          in: query
          description: playerid i.e. 41c44740-d0f6-44ab-8347-3b5d515e5ecf
          required: true
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
...